move files to wechat
