+++ wechat/webservice.py
@@ -11,11 +11,14 @@ import threading
 import time
 
 from mysqlmgr import MysqlMgr
+from mongomgr import MongoManager
 from subprocess import call
 
 from aiohttp.web import Application, Response, StreamResponse, run_app
+import json
 
 mysql_mgr = MysqlMgr(10)
+mongo_mgr = MongoManager()
 
 biz_arr = mysql_mgr.all_biz()
 
@@ -23,12 +26,6 @@ if biz_arr is not None:
     bizs = set(biz_arr)
 else:
     bizs = set([])
-
-running_state_val = {
-    "running" : 0,
-    "need_transaction": 1,
-    "in_transaction" : 2
-}
 
 STATE_RUNNING = 1
 
@@ -39,6 +36,21 @@ running_state = 0
 run_swipe = True
 
 last_history_time = time.clock()
+
+# A thread to save data to database in background
+def insert_to_database(biz, msglist):
+    try:
+        for msg in msglist:
+            print(biz)
+            print(msg['comm_msg_info']['id'])
+            mongo_mgr.enqueue_data(msg['comm_msg_info']['id'], biz, msg )
+    except Exception as e:
+        print(e)
+
+def save_data(biz, msglist_str):
+    save_thread = threading.Thread(target=insert_to_database, args=(biz, msglist_str,))
+    save_thread.setDaemon(True)
+    save_thread.start()
 
 def swipe_for_next_page():
     while run_swipe:
@@ -64,19 +76,32 @@ async def report_msg_home(request):
     data = await request.json()
     global running_state
     global last_history_time
-    print("report_msg_home called")
+    print("report_msg_home: " + data['url'])
+    # print(data['body'])
     last_history_time = time.clock()
-    
+    msglist_str = re.findall('var msgList = \'(\{.*?\})\'\;', data['body'])
+    if len(msglist_str) > 0:
+        msglist_str = msglist_str[0]
+        msglist_str = msglist_str.replace('&quot;','"').replace('\\', '')
+
+        biz_arr = re.findall('__biz=(.*?)\&', data['url'])
+        if len(biz_arr) > 0:
+            biz = biz_arr[0]
+        else:
+            biz = ''
+
+        msglist = json.loads(msglist_str)
+        save_data(biz, msglist['list'])
+
     resp.content_type = 'text/plain'
-    print("report_msg_home called - 0")
     await resp.prepare(request)
     if running_state == STATE_IN_TRANSACTION:
+        # inject javascript code so that it redirects to a new Wechat Account Home
         running_state = STATE_RUNNING
-        resp.write("https://mp.weixin.qq.com/mp/profile_ext?action=home&__biz=" + bizs.pop() + "&scene=124#wechat_redirect")
-    print("report_msg_home called - 1")
-    resp.write("NULL".encode('utf8'))
+        resp.write(("https://mp.weixin.qq.com/mp/profile_ext?action=home&__biz=" + bizs.pop() + "&scene=124#wechat_redirect").encode('utf8'))
+    else:
+        resp.write("NULL".encode('utf8'))
     await resp.write_eof()
-    print("report_msg_home called - 2")
     return resp
 
 async def report_msg_ext(request):
@@ -84,10 +109,19 @@ async def report_msg_ext(request):
     data = await request.json()
     global last_history_time
     print("report_msg_ext called")
+    msg_data = json.loads(data['body'])
+    msglist_str = msg_data['general_msg_list'].replace('\\', '')
+    biz_arr = re.findall('__biz=(.*?)\&', data['url'])
+    if len(biz_arr) > 0:
+        biz = biz_arr[0]
+    else:
+        biz = ''
+    msglist = json.loads(msglist_str)
+    save_data(biz, msglist['list'])
+
     last_history_time = time.clock()
     resp.content_type = 'text/plain'
     await resp.prepare(request)
-    resp.write("NULL".encode('utf8'))
     await resp.write_eof()
     return resp
 
@@ -152,7 +186,7 @@ def start_swipe_thread():
     except Exception:
         print("Error: unable to start thread")
 
-start_swipe_thread()
+# start_swipe_thread()
 loop = asyncio.get_event_loop()
 app = loop.run_until_complete(init(loop))
 run_app(app, host='127.0.0.1', port=9999)
